requirements.txt
aiogram>=3.2.0,<4.0.0
python-dotenv>=1.0.0
pytz>=2023.3